rename the Import directory to import, which is consistent with other packages
